// src/views/song/components/oss.d.ts
declare namespace OSS {
// ...
  type ACLType = "public-read-write" | "public-read" | "private";

  type HTTPMethods = "GET" | "POST" | "DELETE" | "PUT";

  type RedundancyType = "LRS" | "ZRS";

  interface RequestOptions {
    // the operation timeout
    timeout?: number | undefined;
  }
// ...
  interface ObjectMeta {
    /** object name on oss */
    name: string;
    /** object url */
    url: string;
    /** object last modified GMT date, e.g.: 2015-02-19T08:39:44.000Z */
    lastModified: string;
    /** object etag contains ", e.g.: "5B3C1A2E053D763E1B002CC607C5A0FE" */
    etag: string;
    /** object type, e.g.: Normal */
    type: string;
    /** object size, e.g.: 344606 */
    size: number;
    storageClass: StorageType;
    owner?: OwnerType;
  }

  interface BucketPolicy {
    Version: string;
    Statement: Array<{
      Action: string[];
      Effect: "Allow" | "Deny";
      Principal: string[];
      Resource: string[];
    }>;
  }

  interface NormalSuccessResponse {
    /** response status */
    status: number;
    /** response headers */
    /** todo the object in detail */
    headers: object;
    /** response size */
    size: number;
    /**  request total use time (ms) */
    rt: number;
  }

  /**
   * @see x-oss-meta-* in https://help.aliyun.com/document_detail/31978.html for Aliyun user
   * @see x-oss-meta-* in https://www.alibabacloud.com/help/en/doc-detail/31978.html for AlibabaCloud user
   */
  interface UserMeta extends Record<string, string | number> {
    uid: number;
    pid: number;
  }

  interface ObjectCallback {
    /** After a file is uploaded successfully, the OSS sends a callback request to this URL. */
    url: string;
    /** The host header value for initiating callback requests. */
    host?: string | undefined;
    /** The value of the request body when a callback is initiated, for example, key=$(key)&etag=$(etag)&my_var=$(x:my_var). */
    body: string;
    /** The Content-Type of the callback requests initiatiated, It supports application/x-www-form-urlencoded and application/json, and the former is the default value. */
    contentType?: string | undefined;
    customValue?: object | undefined;
    /** extra headers, detail see RFC 2616 */
    headers?: object | undefined;
  }

  interface ModifyData {
    /** object last modified GMT string */
    lastModified: string;
    /** object etag contains ", e.g.: "5B3C1A2E053D763E1B002CC607C5A0FE" */
    etag: string;
  }
// ...
  interface ListObjectsQuery {
    /** search object using prefix key */
    prefix?: string | undefined;
    /** search start from marker, including marker key */
    marker?: string | undefined;
    /** only search current dir, not including subdir */
    delimiter?: string | undefined; // delimiter search scope e.g.
    /** max objects, default is 100, limit to 1000 */
    "max-keys": string | number;
    /** Specifies that the object names in the response are URL-encoded. */
    "encoding-type"?: "url" | "";
  }

  interface ListV2ObjectsQuery {
    /** search object using prefix key */
    prefix?: string;
    /** search start from token, including token key */
    "continuation-token"?: string;
    /** only search current dir, not including subdir */
    delimiter?: string | number;
    /** max objects, default is 100, limit to 1000  */
    "max-keys"?: string;
    /**
     * The name of the object from which the list operation begins.
     * If this parameter is specified, objects whose names are alphabetically greater than the start-after parameter value are returned.
     */
    "start-after"?: string;
    /** Specifies whether to include the information about object owners in the response. */
    "fetch-owner"?: boolean;
    /** Specifies that the object names in the response are URL-encoded. */
    "encoding-type"?: "url" | "";
  }

  interface ListObjectResult {
    objects: ObjectMeta[];
    prefixes: string[];
    isTruncated: boolean;
    nextMarker: string;
    res: NormalSuccessResponse;
  }

  interface PutObjectOptions {
    /** the operation timeout */
    timeout?: number | undefined;
    /** custom mime, will send with Content-Type entity header */
    mime?: string | undefined;
    /** user meta, will send with x-oss-meta- prefix string e.g.: { uid: 123, pid: 110 } */
    meta?: UserMeta | undefined;
    callback?: ObjectCallback | undefined;
    headers?: object | undefined;
  }

  interface PutObjectResult {
    name: string;
    url: string;
    data: object;
    res: NormalSuccessResponse;
  }

  interface PutStreamOptions {
    /** the stream length, chunked encoding will be used if absent */
    contentLength?: number | undefined;
    /** the operation timeout */
    timeout: number;
    /** custom mime, will send with Content-Type entity header */
    mime: string;
    meta: UserMeta;
    callback: ObjectCallback;
    headers?: object | undefined;
  }
// ...
  interface HeadObjectOptions {
    timeout?: number | undefined;
    headers?: object | undefined;
  }

  interface HeadObjectResult {
    /** response status, maybe 200 or 304 */
    status: number;
    meta: UserMeta;
    res: NormalSuccessResponse;
  }

  interface GetObjectOptions {
    timeout?: number | undefined;
    /** The Content-Type of the callback requests initiatiated, It supports application/x-www-form-urlencoded and application/json, and the former is the default value. */
    process?: string | undefined;
    headers?: object | undefined;
  }

  interface GetObjectResult {
    /** file content buffer if file parameter is null or ignore */
    content?: any;
    res: NormalSuccessResponse;
  }

  interface GetStreamOptions {
    timeout?: number | undefined;
    /** The Content-Type of the callback requests initiatiated, It supports application/x-www-form-urlencoded and application/json, and the former is the default value. */
    process?: string | undefined;
    headers?: object | undefined;
  }

  interface GetStreamResult {
    /** readable stream instance if response status is not 200, stream will be null. */
    stream?: any;
    res: NormalSuccessResponse;
  }

  interface CopyObjectOptions {
    timeout?: number | undefined;
    meta?: UserMeta | undefined;
    headers?: object | undefined;
  }

  interface CopyAndPutMetaResult {
    data: ModifyData;
    res: NormalSuccessResponse;
  }

  interface DeleteResult {
    res: NormalSuccessResponse;
  }

  interface DeleteMultiOptions {
    /** quite mode or verbose mode, default is false */
    quiet?: boolean | undefined;
    timeout?: number | undefined;
  }

  interface DeleteMultiResult {
    /** deleted object names list */
    deleted?: string[] | undefined;
    res: NormalSuccessResponse;
  }

  interface ResponseHeaderType {
    "content-type"?: string | undefined;
    "content-disposition"?: string | undefined;
    "cache-control"?: string | undefined;
  }

  interface SignatureUrlOptions {
    /** after expires seconds, the url will become invalid, default is 1800 */
    expires?: number | undefined;
    /** the HTTP method, default is 'GET' */
    method?: HTTPMethods | undefined;
    /** set the request content type */
    "Content-Type"?: string | undefined;
    /**  image process params, will send with x-oss-process e.g.: {process: 'image/resize,w_200'} */
    process?: string | undefined;
    /** traffic limit, range: 819200~838860800 */
    trafficLimit?: number | undefined;
    /** additional signature parameters in url */
    subResource?: object | undefined;
    /** set the response headers for download */
    response?: ResponseHeaderType | undefined;
    /** set the callback for the operation */
    callback?: ObjectCallback | undefined;
  }
// ...
}

// cluster
declare namespace OSS {
  interface ClusterType {
    host: string;
    accessKeyId: string;
    accessKeySecret: string;
  }

  interface ClusterOptions {
    clusters: ClusterType[];
    schedule?: string | undefined;
  }

  class ClusterClient {
    constructor(options: ClusterOptions);

    list(
      query: ListObjectsQuery | null,
      options: RequestOptions
    ): Promise<ListObjectResult>;

    /**
     * @since 6.12.0
     */
    listV2(
      query: ListV2ObjectsQuery | null,
      options?: RequestOptions
    ): Promise<ListObjectResult>;

    put(
      name: string,
      file: any,
      options?: PutObjectOptions
    ): Promise<PutObjectResult>;

    putStream(
      name: string,
      stream: any,
      options?: PutStreamOptions
    ): Promise<{ name: string; res: NormalSuccessResponse }>;

    head(name: string, options?: HeadObjectOptions): Promise<HeadObjectResult>;

    get(
      name: string,
      file?: any,
      options?: GetObjectOptions
    ): Promise<GetObjectResult>;

    getStream(
      name?: string,
      options?: GetStreamOptions
    ): Promise<GetStreamResult>;

    delete(name: string, options?: RequestOptions): Promise<DeleteResult>;

    copy(
      name: string,
      sourceName: string,
      options?: CopyObjectOptions
    ): Promise<CopyAndPutMetaResult>;

    putMeta(
      name: string,
      meta: UserMeta,
      options: RequestOptions
    ): Promise<CopyAndPutMetaResult>;

    deleteMulti(
      names: string[],
      options?: DeleteMultiOptions
    ): Promise<DeleteMultiResult>;

    signatureUrl(name: string, options?: SignatureUrlOptions): string;

    asyncSignatureUrl(
      name: string,
      options?: SignatureUrlOptions
    ): Promise<string>;

    putACL(
      name: string,
      acl: ACLType,
      options?: RequestOptions
    ): Promise<NormalSuccessResponse>;

    restore(
      name: string,
      options?: RequestOptions
    ): Promise<NormalSuccessResponse>;
  }
}